FineUploader installed and functional
The Pages controller has a new function - upload_image

At the moment this functions soul purpose is to return a json object
that fakes a successful upload for test purposes.

I am worried about code repetition if I carry on the javascript in the
style I have done. Considering a rewrite.